requirements.txt
google-generativeai
pillow
pandas
openpyxl